feat(): adicionado pastas para organizacao de view models e rename de namespaces
