+++ Blog.WebApi/Controllers/V1/CategoryController.cs
@@ -2,6 +2,7 @@ using Blog.Data;
 using Blog.Models;
 using Blog.WebApi.Extensions;
 using Blog.WebApi.ViewModels;
+using Blog.WebApi.ViewModels.Categories;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
