+++ Blog.WebApi/ViewModels/Accounts/LoginViewModel.cs
@@ -1,14 +1,14 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace Blog.WebApi.ViewModels
+namespace Blog.WebApi.ViewModels.Accounts
 {
-  public class RegisterViewModel
+  public class LoginViewModel
   {
-    [Required(ErrorMessage = "O campo 'Name' é obrigatório.")]
-    public string Name { get; set; }
-
     [Required(ErrorMessage = "O campo 'Email' é obrigatório.")]
     [EmailAddress(ErrorMessage = "O Email informado é inválido.")]
     public string Email { get; set; }
+
+    [Required(ErrorMessage = "O campo 'Password' é obrigatório.")]
+    public string Password { get; set; }
   }
 }
